服务费记录和删除提示

+++ WebRoot/js/dtjk/maintenanceRecords.js
@@ -48,7 +48,9 @@
          }
 		//删除
 		  function del(id){
-    		  window.location.href="recordsAction.do?method=delEntity&id="+id+gotoUrl();
+			   if( confirm("确定删除这条电梯维保记录？")){
+    		 	 window.location.href="recordsAction.do?method=delEntity&id="+id+gotoUrl();
+    		  }
          }
 		  
 		 //下载  
@@ -60,12 +62,16 @@
 		  	//		批量删除
 $(".del").click(function() {
 	var ids="";
+	var len=0;
 	$(".wei").each(function() {
 			if($(this).children("i").hasClass("gou")) {
 				ids=ids+$(this).find("i").children().val()+",";
+				len++;
 			}
 		})
-	 window.location.href="recordsAction.do?method=deleteEntity&ids="+ids;
+		 if(len>0 && confirm("确定删除这 "+len+" 条记录？")){
+	 		window.location.href="recordsAction.do?method=deleteEntity&ids="+ids;
+		 }
 })
