+++ WebRoot/js/dtjk/yearly.js
@@ -44,7 +44,9 @@
          }
 		//删除
 		  function del(id){
-    		  window.location.href="inspectionAction.do?method=delEntity&id="+id+gotoUrl();
+			   if( confirm("确定删除这条电梯年检记录？")){
+    		  	window.location.href="inspectionAction.do?method=delEntity&id="+id+gotoUrl();
+    		  }
          }
 		  
 		 //下载  
@@ -56,12 +58,16 @@
 		  	//		批量删除
 $(".del").click(function() {
 	var ids="";
+	var len=0;
 	$(".wei").each(function() {
 			if($(this).children("i").hasClass("gou")) {
 				ids=ids+$(this).find("i").children().val()+",";
+				len++;
 			}
 		})
-	 window.location.href="inspectionAction.do?method=deleteEntity&ids="+ids;
+		 if(len>0 && confirm("确定删除这 "+len+" 条记录？")){
+			 window.location.href="inspectionAction.do?method=deleteEntity&ids="+ids;
+		 }
 })
